+++ packages/ui/src/components/atoms/index.ts
@@ -1,4 +1,3 @@
 export { Button } from './Button';
 export { Input } from './Input';
 export { TextArea } from './TextArea';
-export { ThemeToggle } from './ThemeToggle';
