+++ packages/ui/src/components/molecules/index.ts
@@ -1,3 +1,4 @@
 // Molecules
 export * from './Breadcrumb';
 export * from './Drawer';
+export * from './ThemeToggle';
